requirements.txt
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.8.0
PyQt6>=6.6.0
python-docx>=1.1.0
Pillow>=10.0.0
numpy>=1.24.0
